requirements_dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements_dev.txt requirements_dev.in
#
appnope==0.1.3
    # via ipython
asgiref==3.6.0
    # via
    #   -r ./requirements.txt
    #   django
asttokens==2.2.1
    # via stack-data
attrs==22.2.0
    # via pytest
autoflake==2.0.2
    # via -r requirements_dev.in
azure-core==1.26.3
    # via
    #   -r ./requirements.txt
    #   azure-storage-blob
azure-storage-blob==12.15.0
    # via
    #   -r ./requirements.txt
    #   django-storages
backcall==0.2.0
    # via ipython
black==23.3.0
    # via -r requirements_dev.in
brotli==1.0.9
    # via geventhttpclient
build==0.10.0
    # via pip-tools
certifi==2022.12.7
    # via
    #   -r ./requirements.txt
    #   geventhttpclient
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via
    #   -r ./requirements.txt
    #   cryptography
charset-normalizer==3.1.0
    # via
    #   -r ./requirements.txt
    #   requests
click==8.1.3
    # via
    #   black
    #   flask
    #   pip-tools
configargparse==1.5.3
    # via locust
coreapi==2.3.3
    # via
    #   -r ./requirements.txt
    #   drf-yasg
coreschema==0.0.4
    # via
    #   -r ./requirements.txt
    #   coreapi
    #   drf-yasg
coverage[toml]==7.2.2
    # via pytest-cov
cryptography==40.0.1
    # via
    #   -r ./requirements.txt
    #   azure-storage-blob
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via
    #   -r ./requirements.txt
    #   djangorestframework-xml
    #   odfpy
diff-match-patch==20200713
    # via
    #   -r ./requirements.txt
    #   django-import-export
django==3.2.18
    # via
    #   -r ./requirements.txt
    #   django-datetime-utc
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-import-export
    #   django-leaflet
    #   django-storages
    #   djangorestframework
    #   drf-yasg
    #   model-bakery
django-datetime-utc==1.0.4
    # via -r ./requirements.txt
django-debug-toolbar==3.8.1
    # via -r ./requirements.txt
django-extensions==3.2.1
    # via -r ./requirements.txt
django-filter==23.1
    # via -r ./requirements.txt
django-import-export==3.1.0
    # via -r ./requirements.txt
django-leaflet==0.28.3
    # via -r ./requirements.txt
django-storages[azure]==1.13.2
    # via -r ./requirements.txt
djangorestframework==3.14.0
    # via
    #   -r ./requirements.txt
    #   djangorestframework-csv
    #   djangorestframework-gis
    #   drf-amsterdam
    #   drf-extensions
    #   drf-yasg
djangorestframework-csv==2.1.1
    # via
    #   -r ./requirements.txt
    #   drf-amsterdam
djangorestframework-gis==1.0
    # via -r ./requirements.txt
djangorestframework-xml==2.0.0
    # via
    #   -r ./requirements.txt
    #   drf-amsterdam
drf-amsterdam==0.4.1
    # via -r ./requirements.txt
drf-extensions==0.7.1
    # via
    #   -r ./requirements.txt
    #   drf-amsterdam
drf-yasg==1.21.5
    # via -r ./requirements.txt
et-xmlfile==1.1.0
    # via
    #   -r ./requirements.txt
    #   openpyxl
executing==1.2.0
    # via stack-data
factory-boy==3.2.1
    # via
    #   -r requirements_dev.in
    #   pytest-factoryboy
faker==18.3.1
    # via factory-boy
flask==2.2.3
    # via
    #   flask-basicauth
    #   flask-cors
    #   locust
flask-basicauth==0.2.0
    # via locust
flask-cors==3.0.10
    # via locust
gevent==22.10.2
    # via
    #   geventhttpclient
    #   locust
geventhttpclient==2.0.9
    # via locust
greenlet==2.0.2
    # via gevent
idna==3.4
    # via
    #   -r ./requirements.txt
    #   requests
inflection==0.5.1
    # via
    #   -r ./requirements.txt
    #   drf-yasg
    #   pytest-factoryboy
iniconfig==2.0.0
    # via pytest
ipython==8.11.0
    # via -r requirements_dev.in
isodate==0.6.1
    # via
    #   -r ./requirements.txt
    #   azure-storage-blob
isort==5.12.0
    # via -r requirements_dev.in
itsdangerous==2.1.2
    # via flask
itypes==1.2.0
    # via
    #   -r ./requirements.txt
    #   coreapi
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via
    #   -r ./requirements.txt
    #   coreschema
    #   flask
locust==2.15.1
    # via -r requirements_dev.in
markuppy==1.14
    # via
    #   -r ./requirements.txt
    #   tablib
markupsafe==2.1.2
    # via
    #   -r ./requirements.txt
    #   jinja2
    #   werkzeug
matplotlib-inline==0.1.6
    # via ipython
model-bakery==1.10.1
    # via -r requirements_dev.in
msgpack==1.0.5
    # via locust
mypy-extensions==1.0.0
    # via black
odfpy==1.4.1
    # via
    #   -r ./requirements.txt
    #   tablib
openpyxl==3.1.2
    # via
    #   -r ./requirements.txt
    #   tablib
packaging==23.0
    # via
    #   -r ./requirements.txt
    #   black
    #   build
    #   drf-yasg
    #   pytest
parso==0.8.3
    # via jedi
pathspec==0.11.1
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.4.0
    # via -r ./requirements.txt
pip==23.0.1
    # via pip-tools
pip-tools==6.12.3
    # via -r requirements_dev.in
platformdirs==3.2.0
    # via black
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.38
    # via ipython
psutil==5.9.4
    # via locust
psycopg2-binary==2.9.5
    # via -r ./requirements.txt
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via
    #   -r ./requirements.txt
    #   cffi
pyflakes==3.0.1
    # via autoflake
pygments==2.14.0
    # via ipython
pyproject-hooks==1.0.0
    # via build
pytest==7.2.2
    # via
    #   -r requirements_dev.in
    #   pytest-cov
    #   pytest-django
    #   pytest-factoryboy
pytest-cov==4.0.0
    # via -r requirements_dev.in
pytest-django==4.5.2
    # via -r requirements_dev.in
pytest-factoryboy==2.5.1
    # via -r requirements_dev.in
python-dateutil==2.8.2
    # via
    #   -r ./requirements.txt
    #   django-datetime-utc
    #   faker
    #   time-machine
pytz==2023.3
    # via
    #   -r ./requirements.txt
    #   django
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0
    # via
    #   -r ./requirements.txt
    #   tablib
pyzmq==25.0.2
    # via locust
requests==2.28.2
    # via
    #   -r ./requirements.txt
    #   azure-core
    #   coreapi
    #   locust
roundrobin==0.0.4
    # via locust
ruamel-yaml==0.17.21
    # via
    #   -r ./requirements.txt
    #   drf-yasg
sentry-sdk==1.18.0
    # via -r ./requirements.txt
setuptools==67.6.1
    # via
    #   -r ./requirements.txt
    #   gevent
    #   pip-tools
six==1.16.0
    # via
    #   -r ./requirements.txt
    #   azure-core
    #   djangorestframework-csv
    #   flask-cors
    #   geventhttpclient
    #   isodate
    #   python-dateutil
sqlparse==0.4.3
    # via
    #   -r ./requirements.txt
    #   django
    #   django-debug-toolbar
stack-data==0.6.2
    # via ipython
tablib[html,ods,xls,xlsx,yaml]==3.4.0
    # via
    #   -r ./requirements.txt
    #   django-import-export
time-machine==2.9.0
    # via -r requirements_dev.in
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.5.0
    # via
    #   -r ./requirements.txt
    #   azure-core
    #   azure-storage-blob
    #   locust
    #   pytest-factoryboy
unicodecsv==0.14.1
    # via
    #   -r ./requirements.txt
    #   djangorestframework-csv
uritemplate==4.1.1
    # via
    #   -r ./requirements.txt
    #   coreapi
    #   drf-yasg
urllib3==1.26.15
    # via
    #   -r ./requirements.txt
    #   requests
    #   sentry-sdk
wcwidth==0.2.6
    # via prompt-toolkit
werkzeug==2.2.3
    # via
    #   flask
    #   locust
wheel==0.40.0
    # via pip-tools
xlrd==2.0.1
    # via
    #   -r ./requirements.txt
    #   tablib
xlwt==1.3.0
    # via
    #   -r ./requirements.txt
    #   tablib
zope-event==4.6
    # via gevent
zope-interface==6.0
    # via gevent
